+++ src/app/[locale]/marketplace/catalogConfiguration.ts
