+++ src/app/[locale]/marketplace/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import { useTranslations } from 'next-intl';
 import ContentPasteSearchIcon from '@mui/icons-material/ContentPasteSearch';
 import ListHeader from 'components/basics/ListHeader';
-import { CatalogConfiguration } from './catalogConfiguration';
+import { CatalogConfiguration } from 'app/[locale]/marketplace/catalogConfiguration';
 import Image from 'next/image';
 import { Breadcrumbs } from 'components/basics/Breadcrumbs';
 import { getCatalogBreadcrumbs } from 'app/catalog/breadcrumbs';
@@ -61,18 +61,18 @@ export default function Page() {
                 <Box marginBottom="1em">
                     <Breadcrumbs links={breadcrumbLinks} />
                 </Box>
-                    <Box display="flex" alignItems="center" justifyContent="space-between" mb={2}>
-                        <ListHeader header={t('marketplaceTitle')} subHeader={t('marketplaceSubtitle')} />
-                        <Button
-                            variant="contained"
-                            startIcon={<InsightsIcon />}
-                            onClick={() => setIsConstructionDialogOpen(true)}
-                            sx={{ whiteSpace: 'nowrap', ml: 2 }}
-                            aria-label="Chatbot"
-                        >
-                            Chatbot
-                        </Button>
-                    </Box>
+                <Box display="flex" alignItems="center" justifyContent="space-between" mb={2}>
+                    <ListHeader header={t('marketplaceTitle')} subHeader={t('marketplaceSubtitle')} />
+                    <Button
+                        variant="contained"
+                        startIcon={<InsightsIcon />}
+                        onClick={() => setIsConstructionDialogOpen(true)}
+                        sx={{ whiteSpace: 'nowrap', ml: 2 }}
+                        aria-label="Chatbot"
+                    >
+                        Chatbot
+                    </Button>
+                </Box>
                 <Box display="flex" alignItems="center" marginTop="1.5rem" marginBottom="1.5rem">
                     <ContentPasteSearchIcon fontSize="large" color="primary" sx={{ mr: 1 }} />
                     <Typography variant="h4" fontWeight={600} component="h2">
@@ -111,7 +111,7 @@ export default function Page() {
                                     {t('articleCount', { count: config.articleCount })}
                                 </Typography>
                                 <IconButton
-                                    onClick={() => navigate.push(`/catalog/${encodeURIComponent(manufacturer)}`)}
+                                    onClick={() => navigate.push(`/marketplace/catalog?manufacturer=${encodeURIComponent(manufacturer)}`)}
                                     sx={{ bgcolor: theme.palette.primary.light, color: theme.palette.primary.contrastText, '&:hover': { bgcolor: theme.palette.primary.main }, cursor: 'pointer' }}
                                 >
                                     <ArrowForwardIcon />
